feat(updates): terminal notifications, and channel changes that can be taken back
Notifications
-------------
"Update available" is posted at the point of discovery, inside checkForUpdates,
rather than by collecting newUpdateEvents. That flow is a single-consumer
channel, so a second collector would steal events from the shell instead of
duplicating them — and any process-wide collector would have to start from
Application.onCreate to exist at all on the run that matters, a periodic check
with the app never opened. The discovery site also already owns the "is this
worth announcing" gate, so the notification and the snackbar agree by
construction.

Each channel notifies into its own slot: the two checks run concurrently, and a
shared slot means whichever finishes second silently erases the other. Neither
posts when that channel auto-downloads (the download is the announcement) or for
the build already running — the nightly channel needs the latter, since its
freshness check compares against what the store last saw and a fresh install
reads the running nightly as new.

On success the worker cancels the progress notification and posts "ready to
install" into a slot of its own, since WorkManager tears the progress one down
when the worker ends and there is nothing left to morph. The tap routes through
the activity into installApk rather than straight at the system installer, so a
dismissed installer dialog is still detected. The offer survives its own tap and
does not re-alert, because declining leaves a perfectly good APK behind.
Installing replaces the process before anything can clean up, so the first start
afterwards is what takes the notification down.

Channel changes
---------------
Switching to nightly now mirrors restoring to stable: flip the switch, then
offer the build the switch was for on a sheet that can put it back. The two
carry their undo with them, so asking for the wording without an escape route
is not expressible.

The sheet no longer blocks leaving mid-download. That guard predates
foreground-service downloads, when leaving really would have killed one; it now
only traps whoever is on a slow connection. For a channel change, leaving *is*
declining it, so the back gesture undoes it exactly as the button does — and
undoing has to cancel the work, not just reset the state, or the download
finishes anyway and goes on offering itself.

A sheet open when the process dies is restored on the next launch. The shell
reads that slot once at startup rather than observing it: both hosts write it as
they open, and a live collector would answer by stacking a second copy on top.

startDownload now reports whether the request was taken. WorkManager's KEEP
policy drops a duplicate silently, and a sheet opened for a release that was
never requested renders the other download's progress and then offers to install
its APK under this release's name.

+++ android/app/src/main/java/it/attendance100/mybicocca/data/notification/AppNotifier.kt
@@ -145,8 +145,19 @@ class AppNotifier @Inject constructor(
     }
 
     private companion object {
+        /**
+         * Distinct per action *within one notification*, which is what the actions on a single
+         * notification need: without it, PendingIntent equality collapses two buttons into
+         * whichever was created first.
+         *
+         * Across notifications it is a hash, not an injection - the Long arithmetic keeps the
+         * result stable and non-negative, but squeezing (id, index) into an Int cannot be
+         * collision-free, and multiplying by 100 loses two more bits. Two entity notifications
+         * whose ids differ by exactly 2^29 would share request codes. Live entity notifications
+         * number in the dozens, so that is a theoretical loss rather than one to design around.
+         */
         fun actionRequestCode(id: NotificationId, index: Int): Int =
-            (id.value * 100 + index) and Int.MAX_VALUE
+            ((id.value.toLong() * 100L + index) and Int.MAX_VALUE.toLong()).toInt()
 
         /** Summaries live in the entity id space, keyed by group, so they can't hit a real slot. */
         fun summaryIdFor(groupKey: String): Int = NotificationId.idFor("group-summary", groupKey)
